STYLE: Remove `static_cast<unsigned int>` from dimension constants
Following ITK pull request https://github.com/InsightSoftwareConsortium/ITK/pull/5320 commit https://github.com/InsightSoftwareConsortium/ITK/commit/9e533c4bb5ac99a7edb72b2f7e39e8c00d69564d

+++ Common/OpenCL/Filters/itkGPUResampleImageFilter.hxx
@@ -482,7 +482,7 @@ GPUResampleImageFilter<TInputImage, TOutputImage, TInterpolatorPrecisionType, TT
   }
 
   unsigned int mem_size_DF = 0;
-  switch (static_cast<unsigned int>(OutputImageDimension))
+  switch (OutputImageDimension)
   {
     case 1:
       mem_size_DF = totalDFSize * sizeof(cl_float);
@@ -547,7 +547,7 @@ GPUResampleImageFilter<TInputImage, TOutputImage, TInterpolatorPrecisionType, TT
     // The deformation field size is the second argument in the
     // pre/loop/post kernel, i.e. index is 1.
     const cl_uint dfSizeKernelIndex = 1;
-    switch (static_cast<unsigned int>(OutputImageDimension))
+    switch (OutputImageDimension)
     {
       case 1:
       {
